requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
PyJWT
openai
PyMuPDF
python-multipart